TempFile, TempDir: Drop handle, add open method
Drops the `file` or `dir` handle from the `Temp*` types,
replacing them with an `open` method on each.

This puts the responsibility of opening and closing the file or
directory handle on the caller, ensuring that we're not double-closing a
file that the user already closed.

To prevent confusion between Dir.close and TempDir.close,
the destructors have been renamed to `deinit`.

Resolves #2